AI, Pathfinding, & Ducking
Implemented the A* Pathfinding Project extension and modified a metric
butt ton of code to go with it for the player. Player moves along the
rail, can be stopped, enemies can stop and shoot while following their
own rails, ect.

Currently trying to work out the duck mechanic. It partially works but
there is a struggle between switching in and out rapidly in the code.
Not to self: put the endDuck bool in the movement code, test further.

Added EnemyManager for some bool logic to manage enemies. CameraBehavior
does nothing at the moment. PathEvent is also bool logic to manage the
paths for each event that gets scripted.

NOTE: Bullets go through walls. What? Why? Need to fix the Unity
bulletFired prefab - not in code.