EnemyAI, Camera, Event & SpawnWaves
Created enemyAI, spawning enemy waves, camera control for the player.
Lots of scripting and fine tuned ducking as well. Largest commit I
believe as of now.

So far, the game works as how I want it to. The next step is to just add
some nextNode and nextEvent code and I should be done.